fix(auth): clear oauth bootstrap ids on session clear
Security review flagged that the cached bootstrap ids outlived the OAuth session.
Move the id cache into oauthClient (alongside the session) and null it in clearSession()
so a later login in the same tab can't read the previous account's ids. getAppContext
reads them via getOAuthContextIds() and stays a leaf module (no cycle).
